Added four tasks for Lesson3. Added classes.
