+++ HelloWorld/HelloWorld.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31624.102
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelloWorld", "HelloWorld\HelloWorld.csproj", "{926B93FA-B73B-470A-9118-B5300CE14D6D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TMS_Hometasks", "HelloWorld\TMS_Hometasks.csproj", "{926B93FA-B73B-470A-9118-B5300CE14D6D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
