+++ HelloWorld/HelloWorld/Hometask_Lesson1.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TMS_Hometasks
+{
+    class Hometask_Lesson1
+    {
+        public static void Lesson1_Task1()
+        {
+            Console.WriteLine("Hello world!");
+        }
+
+        public static void Lesson1_Task2()
+        {
+            Console.WriteLine("Enter your name.");
+            string user = Console.ReadLine();
+
+            Console.WriteLine($"Hello {user}");
+        }
+    }
+}
