+++ HelloWorld/HelloWorld/bin/Debug/net5.0/TMS_Hometasks.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v5.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v5.0": {
+      "TMS_Hometasks/1.0.0": {
+        "runtime": {
+          "TMS_Hometasks.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "TMS_Hometasks/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
