+++ HelloWorld/HelloWorld/bin/Debug/net5.0/TMS_Hometasks.runtimeconfig.dev.json
@@ -1,0 +1,9 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "C:\\Users\\kgrebenyuk\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\kgrebenyuk\\.nuget\\packages",
+      "C:\\Program Files (x86)\\Microsoft Visual Studio\\Shared\\NuGetPackages"
+    ]
+  }
+}
